// src/api/openapi/auto-generated/models/LossAssetAttributes.ts
export interface LossAssetAttributes {
    /**
     * 
     * @type {string}
     * @memberof LossAssetAttributes
     */
    occupancyScheme?: string;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    occupancyCode?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    numberOfStories?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    basementCode?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    totalValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    buildingValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    contentsValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    inventoryValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    downtimeValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    firstFloorElevation?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    floorAreaSqm?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    electricityCostUsd?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    electricityCost?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    coolingSystemProbability?: number;
    /**
     * 
     * @type {WorkIntensity}
     * @memberof LossAssetAttributes
     */
    workIntensity?: WorkIntensity;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    remoteWorkRatio?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    waterConsumption?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    waterShadowPriceRatio?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    annualGrowth?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    annualVolatility?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    discountRate?: number;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    financialBaseYear?: number;
    /**
     * 
     * @type {string}
     * @memberof LossAssetAttributes
     */
    subIndustryCode?: string;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    salesMargin?: number;
    /**
     * 
     * @type {CurrencyCode}
     * @memberof LossAssetAttributes
     */
    valueCurrencyCode?: CurrencyCode;
    /**
     * 
     * @type {number}
     * @memberof LossAssetAttributes
     */
    informationFactor?: number;
    /**
     * 
     * @type {WindowPane}
     * @memberof LossAssetAttributes
     */
    windowPane?: WindowPane;
    /**
     * 
     * @type {VentType}
     * @memberof LossAssetAttributes
     */
    ventType?: VentType;
    /**
     * 
     * @type {RoofCover}
     * @memberof LossAssetAttributes
     */
    roofCover?: RoofCover;
}