// src/api/openapi/auto-generated/models/RoofCover.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const RoofCover = {
    wood: 'wood',
    other: 'other',
    unknown: 'unknown',
    unknownDefaultOpenApi: '11184809'
} as const;
export type RoofCover = typeof RoofCover[keyof typeof RoofCover];


export function instanceOfRoofCover(value: any): boolean {
    for (const key in RoofCover) {
        if (Object.prototype.hasOwnProperty.call(RoofCover, key)) {
            if ((RoofCover as Record<string, RoofCover>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RoofCoverFromJSON(json: any): RoofCover {
    return RoofCoverFromJSONTyped(json, false);
}

export function RoofCoverFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoofCover {
    return json as RoofCover;
}

export function RoofCoverToJSON(value?: RoofCover | null): any {
    return value as any;
}