// src/api/openapi/auto-generated/models/VentType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const VentType = {
    exposed: 'exposed',
    nonExposed: 'nonExposed',
    none: 'none',
    unknown: 'unknown',
    unknownDefaultOpenApi: '11184809'
} as const;
export type VentType = typeof VentType[keyof typeof VentType];


export function instanceOfVentType(value: any): boolean {
    for (const key in VentType) {
        if (Object.prototype.hasOwnProperty.call(VentType, key)) {
            if ((VentType as Record<string, VentType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VentTypeFromJSON(json: any): VentType {
    return VentTypeFromJSONTyped(json, false);
}

export function VentTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): VentType {
    return json as VentType;
}

export function VentTypeToJSON(value?: VentType | null): any {
    return value as any;
}