// src/api/openapi/auto-generated/models/WindowPane.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const WindowPane = {
    single: 'single',
    multiple: 'multiple',
    none: 'none',
    unknown: 'unknown',
    unknownDefaultOpenApi: '11184809'
} as const;
export type WindowPane = typeof WindowPane[keyof typeof WindowPane];


export function instanceOfWindowPane(value: any): boolean {
    for (const key in WindowPane) {
        if (Object.prototype.hasOwnProperty.call(WindowPane, key)) {
            if ((WindowPane as Record<string, WindowPane>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WindowPaneFromJSON(json: any): WindowPane {
    return WindowPaneFromJSONTyped(json, false);
}

export function WindowPaneFromJSONTyped(json: any, ignoreDiscriminator: boolean): WindowPane {
    return json as WindowPane;
}

export function WindowPaneToJSON(value?: WindowPane | null): any {
    return value as any;
}